+++ src/Modifiers/Order/CreateOrFetchOrderModel.php
@@ -28,6 +28,11 @@ class CreateOrFetchOrderModel extends AbstractModifier
         );
     }
 
+    public function supports(mixed $entity): bool
+    {
+        return $entity instanceof ExternalOrderInterface;
+    }
+
     /**
      * @param ExternalOrderInterface $model
      * @param OrderInterface|null    $result